+++ src/storage_collision/SolidityByExample/HackMe.sol
@@ -4,14 +4,14 @@ pragma solidity ^0.8.13;
 contract HackMe {
     address public lib;
     address public owner;
-    uint public someNumber;
+    uint256 public someNumber;
 
     constructor(address _lib) {
         lib = _lib;
         owner = msg.sender;
     }
 
-    function doSomething(uint _num) public {
+    function doSomething(uint256 _num) public {
         lib.delegatecall(abi.encodeWithSignature("doSomething(uint256)", _num));
     }
 }
