+++ src/storage_collision/SolidityByExample/HackMeFixed.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.13;
 
 contract HackMeFixed {
-    uint public someNumber; // Moving this variable to the first storage slot instead of the 3rd fixes the issue
+    uint256 public someNumber; // Moving this variable to the first storage slot instead of the 3rd fixes the issue
     address public lib;
     address public owner;
 
@@ -11,7 +11,7 @@ contract HackMeFixed {
         owner = msg.sender;
     }
 
-    function doSomething(uint _num) public {
+    function doSomething(uint256 _num) public {
         lib.delegatecall(abi.encodeWithSignature("doSomething(uint256)", _num));
     }
 }
